Rename the verify message to follow naming convetions
