+++ src/components/screens/verifyMessage/result.js
